Refactor: FooterBottomLayout의 이름을 FooterBottomContainer로 변경 - Page의 child 컴포넌트로, 그에 맞는 이름인 Container로 변경
