+++ src/Footer/FooterBottomLayout/FooterBottomContainer.test.jsx
@@ -1,16 +1,16 @@
 import { render } from '@testing-library/react';
 
-import FooterBottomLayout from './FooterBottomLayout';
+import FooterBottomContainer from './FooterBottomContainer';
 
 describe('FooterBottomContainer', () => {
   it('renders the title', () => {
-    const { container } = render(<FooterBottomLayout />);
+    const { container } = render(<FooterBottomContainer />);
 
     expect(container).toHaveTextContent('2022');
   });
 
   it('renders the chair icon', () => {
-    const { getByTitle } = render(<FooterBottomLayout />);
+    const { getByTitle } = render(<FooterBottomContainer />);
 
     expect(getByTitle('chair')).not.toBeNull();
   });
